Refactor / cleanup CRUD example.
