+++ examples/crud-mcp/wrangler.jsonc
@@ -2,7 +2,6 @@
   "name": "crud-mcp",
   "main": "src/index.ts",
   "compatibility_date": "2025-03-10",
-  "compatibility_flags": ["nodejs_compat"],
   "durable_objects": {
     "bindings": [
       {
