+++ package.json
@@ -12,15 +12,5 @@
   ],
   "scripts": {
     "test": "yarn workspaces run test"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.9.0",
-    "mcp-server": "^0.0.9",
-    "sqlite": "^5.1.1",
-    "sqlite3": "^5.1.7",
-    "uuid": "^11.1.0"
-  },
-  "build": {
-    "script": "tsc --declaration --outDir dist --emitDeclarationOnly && cp src/**/*.ts dist/"
   }
 }
